The cpu frequency is stored in a 32-bit not a 16-bit int.
Reported by:	Meno Abels <meno dot abels at adviser dot com>

+++ monitor-core/srclib/libmetrics/freebsd/metrics.c
@@ -195,7 +195,7 @@ cpu_speed_func ( void )
    }
 
 done:
-   val.uint16 = freq;
+   val.uint32 = freq;
 
    return val;
 }
